base,general: move CancelParams to base.go
