+++ base_json_test.go
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: Copyright 2021 The Go Language Server Authors
+// SPDX-License-Identifier: BSD-3-Clause
+
+//go:build !gojay
+// +build !gojay
+
+package protocol
+
+import (
+	"encoding/json"
+	"testing"
+)
+
+func TestCancelParams(t *testing.T) {
+	testCancelParams(t, json.Marshal, json.Unmarshal)
+}
